Add RegularWearable
